Make sample drawing always appear to move from right to left
Previously when there were fewer samples than would fill the screen,
the drawing would begin at the left, then once the full width of the
drawing area was filled, the sample window would move and at that point
appear to be moving from right to left.

This makes drawing *always* appear to be moving from right to left,
even when there are fewer samples than needed to fill the drawing area.

For consistency, this also updates the `shouldDrawSilencePadding` case
to appear to move from right to left.