fix silly spelling mistake in public API
Unfortunately this a backwards incompatible change to the public API,
so this will go out as a new major release. I decided against @deprecating it
since it's in too many places and too confusing with two almost identical
spellings. Plus it should mostly be an easy Find+Replace fix to migrate.

fixes #64